feat: Set up data module
